requirements.txt
torch>=2.0.0
torchvision>=0.16.0
numpy>=1.25.0
Pillow>=10.0.0
matplotlib>=3.8.0
